feat: apply theme-color meta-data for navigation bar color change
